refactor(ui): extract chat rendering and console helpers
Move chat message rendering, markdown output, agent activity, and diff review into focused components. Split dashboard and chat helper logic out of useRuntimeConsole so the hook keeps orchestration state while helper modules carry pure transformations and snapshot refresh logic.
