+++ ui/src/app/useRuntimeConsole.ts
@@ -16,15 +16,9 @@ import {
   getBudget,
   getChatSession,
   getChatSessions,
-  getControlPlaneConfig,
-  getHealth,
   getModels,
-  getProviderPresets,
   getProviders,
   getRequestLedger,
-  getRetentionRuns,
-  getRuntimeStats,
-  getSession,
   setProviderAPIKey as setProviderAPIKeyRequest,
   upsertPricebookEntry as upsertPricebookEntryRequest,
   deletePricebookEntry as deletePricebookEntryRequest,
@@ -51,15 +45,26 @@ import {
   createAgentChatSession as createAgentChatSessionRequest,
   deleteAgentChatSession as deleteAgentChatSessionRequest,
   deleteProvider as deleteProviderRequest,
-  getAgentAdapters,
   getAgentChatSession,
-  getAgentChatSessions,
   streamAgentChatSession,
   setProviderBaseURL as setProviderBaseURLRequest,
   setProviderName as setProviderNameRequest,
   setProviderCustomName as setProviderCustomNameRequest,
 } from "../lib/api";
 import type { PolicyRuleUpsertPayload, ResolveAgentChatApprovalPayload, AgentChatGrantFilter } from "../lib/api";
+import {
+  approvalRecordToPending,
+  buildAssistantToolCallMessage,
+  buildMessagesForSubmission,
+  buildSyntheticChatResult,
+  defaultModelForProvider,
+  deriveChatSessionTitle,
+  humanizeChatError,
+  isModelValidForProvider,
+  renderAgentChatSessionSummary,
+  renderChatSessionSummary,
+} from "./runtimeConsoleChatHelpers";
+import { deriveSessionState, resolveDashboardSnapshot } from "./runtimeConsoleDashboard";
 import type {
   BudgetStatusResponse,
   AccountSummaryResponse,
@@ -88,14 +93,8 @@ import type {
   RuntimeHeaders,
   SessionResponse,
   RetentionRunData,
-  RuntimeStatsResponse,
 } from "../types/runtime";
 
-// Single-user mode: the session shape is a fixed label. Kept around so
-// the status bar has something to show without bespoke wiring.
-type SessionState = {
-  label: string;
-};
 type NoticeState = {
   kind: "success" | "error";
   message: string;
@@ -107,6 +106,8 @@ function readLocalStorage(key: string): string {
   }
   return window.localStorage.getItem(key) ?? "";
 }
+
+export { humanizeChatError };
 
 export function useRuntimeConsole() {
   const [health, setHealth] = useState<HealthResponse | null>(null);
@@ -637,7 +638,7 @@ export function useRuntimeConsole() {
         setBudget(scopedBudget.data);
       } catch {
         // Best-effort; the gateway may not have a per-provider budget
-        // record for this slice yet.
+        // record for this provider.
       }
 
       await refreshChatSessionState(sessionID);
@@ -1616,492 +1617,4 @@ export function useRuntimeConsole() {
   };
 }
 
-// humanizeChatError translates raw gateway/provider errors into something
-// an operator can act on. The backend's "api key is required for cloud
-// provider X when stub mode is disabled" carries internal vocabulary
-// (stub mode) that's noise to the user — they just need to know they
-// should add a key. Falls back to the raw string when no pattern matches.
-export function humanizeChatError(raw: string): string {
-  const apiKeyPattern = /api key is required for cloud provider (\S+)/i;
-  const m = raw.match(apiKeyPattern);
-  if (m) {
-    return `${m[1]} has no API key. Open the Providers tab and add one.`;
-  }
-  return raw;
-}
-
-function deriveChatSessionTitle(message: string): string {
-  const normalized = message.trim().replace(/\s+/g, " ");
-  if (!normalized) {
-    return "New chat";
-  }
-  if (normalized.length <= 48) {
-    return normalized;
-  }
-  return `${normalized.slice(0, 45)}...`;
-}
-
-function buildMessagesForSubmission(activeSession: ChatSessionRecord | null, message: string, systemPrompt = ""): ChatMessage[] {
-  // Replay is now a near-trivial transform: the persisted message
-  // stream is already in submission order. We carry content_blocks
-  // and tool_error through verbatim so Anthropic-aware history
-  // (thinking blocks, failed tool results) survives cross-provider
-  // resubmission.
-  const history: ChatMessage[] = (activeSession?.messages ?? [])
-    .filter((m) => m.id && !m.id.startsWith("pending-"))
-    .map((m) => persistedMessageToChatMessage(m));
-  const prefix: ChatMessage[] = systemPrompt.trim() ? [{ role: "system", content: systemPrompt.trim() }] : [];
-  return [...prefix, ...history, { role: "user", content: message }];
-}
-
-function persistedMessageToChatMessage(m: ChatSessionRecord["messages"] extends (infer U)[] | undefined ? U : never): ChatMessage {
-  const ext = {
-    ...(m.content_blocks ? { content_blocks: m.content_blocks } : {}),
-    ...(m.tool_error ? { tool_error: m.tool_error } : {}),
-  };
-  if (m.role === "assistant") {
-    return {
-      role: "assistant",
-      content: m.content,
-      ...(m.tool_calls && m.tool_calls.length > 0 ? { tool_calls: m.tool_calls } : {}),
-      ...ext,
-    } as ChatMessage;
-  }
-  if (m.role === "tool") {
-    return {
-      role: "tool",
-      content: m.content ?? "",
-      tool_call_id: m.tool_call_id ?? "",
-      ...ext,
-    } as ChatMessage;
-  }
-  // user / system / unknown
-  return {
-    role: m.role === "system" ? "system" : "user",
-    content: m.content ?? "",
-    ...ext,
-  } as ChatMessage;
-}
-
-function buildAssistantToolCallMessage(
-  content: string,
-  toolCalls: Array<{ id: string; name: string; arguments: string }>,
-): ChatMessage {
-  return {
-    role: "assistant",
-    content: content || null,
-    tool_calls: toolCalls.map((tc) => ({
-      id: tc.id,
-      type: "function",
-      function: { name: tc.name, arguments: tc.arguments },
-    })),
-  };
-}
-
-function buildSyntheticChatResult(headers: RuntimeHeaders, selectedModel: string, content: string): ChatResponse {
-  return {
-    id: headers.requestId || "stream",
-    model: headers.resolvedModel || selectedModel,
-    choices: [{ index: 0, message: { role: "assistant", content }, finish_reason: "stop" }],
-    usage: { prompt_tokens: 0, completion_tokens: 0, total_tokens: 0 },
-  };
-}
-
-function defaultModelForProvider(provider: ProviderFilter, models: ModelResponse["data"], providers: ProviderStatusResponse["data"], presets: ProviderPresetRecord[]): string {
-  if (provider === "auto") {
-    return "";
-  }
-
-	const providerRecord = providers.find((entry) => entry.name === provider);
-	const scopedModels = models.filter((entry) => entry.metadata?.provider === provider);
-	const preset = presets.find((entry) => entry.id === provider);
-	if (providerRecord?.default_model) {
-		return providerRecord.default_model;
-	}
-
-	if (providerRecord) {
-		return scopedModels.find((entry) => entry.metadata?.default)?.id ?? scopedModels[0]?.id ?? providerRecord.models?.[0] ?? "";
-	}
-
-	return scopedModels.find((entry) => entry.metadata?.default)?.id ?? scopedModels[0]?.id ?? preset?.default_model ?? "";
-}
-
-function isModelValidForProvider(model: string, provider: ProviderFilter, models: ModelResponse["data"], providers: ProviderStatusResponse["data"], presets: ProviderPresetRecord[]): boolean {
-  if (!model || provider === "auto") {
-    return true;
-  }
-
-  if (models.some((entry) => entry.id === model && entry.metadata?.provider === provider)) {
-    return true;
-  }
-
-	const providerRecord = providers.find((entry) => entry.name === provider);
-	if (providerRecord?.default_model === model || providerRecord?.models?.includes(model)) {
-		return true;
-	}
-	if (providerRecord) {
-		return false;
-	}
-
-	const preset = presets.find((entry) => entry.id === provider);
-	return preset?.default_model === model;
-}
-
-function renderChatSessionSummary(session: ChatSessionRecord): ChatSessionsResponse["data"][number] {
-  const messages = session.messages ?? [];
-  const calls = session.provider_calls ?? [];
-  const lastCall = calls[calls.length - 1];
-  return {
-    id: session.id,
-    title: session.title,
-    message_count: messages.length,
-    provider_call_count: calls.length,
-    created_at: session.created_at,
-    updated_at: session.updated_at,
-    last_model: lastCall?.model,
-    last_provider: lastCall?.provider,
-    last_cost_usd: lastCall?.cost_usd,
-    last_request_id: lastCall?.request_id,
-  };
-}
-
-// approvalRecordToPending projects a full approval row down to
-// the banner-essentials shape. Used on the initial GET-list refetch so
-// rows from refetch and rows from streamed `approval.requested` events
-// share one storage shape.
-function approvalRecordToPending(row: AgentChatApprovalRecord): PendingAgentApproval {
-  return {
-    approval_id: row.id,
-    session_id: row.session_id,
-    adapter_id: row.adapter_id,
-    tool_kind: row.tool_kind,
-    tool_name: row.tool_name,
-    scope_choices: row.scope_choices,
-    created_at: row.created_at,
-    expires_at: row.expires_at,
-  };
-}
-
-function renderAgentChatSessionSummary(session: AgentChatSessionRecord): AgentChatSessionsResponse["data"][number] {
-  return {
-    id: session.id,
-    title: session.title,
-    adapter_id: session.adapter_id,
-    driver_kind: session.driver_kind,
-    native_session_id: session.native_session_id,
-    workspace: session.workspace,
-    workspace_branch: session.workspace_branch,
-    status: session.status,
-    message_count: session.messages?.length ?? 0,
-    created_at: session.created_at,
-    updated_at: session.updated_at,
-  };
-}
-
 export type RuntimeConsoleViewModel = ReturnType<typeof useRuntimeConsole>;
-
-type DashboardResults = {
-  health: PromiseSettledResult<HealthResponse>;
-  session: PromiseSettledResult<SessionResponse>;
-  models: PromiseSettledResult<ModelResponse>;
-  providers: PromiseSettledResult<ProviderStatusResponse>;
-  providerPresets: PromiseSettledResult<{ object: string; data: ProviderPresetRecord[] }>;
-  agentAdapters: PromiseSettledResult<{ object: string; data: AgentAdapterRecord[] }>;
-  budget: PromiseSettledResult<BudgetStatusResponse>;
-  accountSummary: PromiseSettledResult<AccountSummaryResponse>;
-  chatSessions: PromiseSettledResult<ChatSessionsResponse>;
-  agentChatSessions: PromiseSettledResult<AgentChatSessionsResponse>;
-  requestLedger: PromiseSettledResult<RequestLedgerResponse>;
-  controlPlaneConfig: PromiseSettledResult<ConfiguredStateResponse>;
-  retentionRuns: PromiseSettledResult<{ object: string; data: RetentionRunData[] }>;
-  runtimeStats: PromiseSettledResult<RuntimeStatsResponse>;
-};
-
-type DashboardPreviousState = {
-  providers: ProviderStatusResponse["data"];
-  agentAdapters: AgentAdapterRecord[];
-  budget: BudgetStatusResponse["data"] | null;
-  accountSummary: AccountSummaryResponse["data"] | null;
-  chatSessions: ChatSessionsResponse["data"];
-  activeChatSession: ChatSessionRecord | null;
-  agentChatSessions: AgentChatSessionsResponse["data"];
-  activeAgentChatSession: AgentChatSessionRecord | null;
-  requestLedger: RequestLedgerResponse["data"];
-  controlPlaneConfig: ConfiguredStateResponse["data"] | null;
-  retentionRuns: RetentionRunData[];
-  retentionLastRun: RetentionRunData | null;
-};
-
-type DashboardSnapshot = {
-  health: HealthResponse;
-  sessionInfo: SessionResponse["data"] | null;
-  models: ModelResponse["data"];
-  providers: ProviderStatusResponse["data"];
-  providerPresets: ProviderPresetRecord[];
-  agentAdapters: AgentAdapterRecord[];
-  budget: BudgetStatusResponse["data"] | null;
-  accountSummary: AccountSummaryResponse["data"] | null;
-  chatSessions: ChatSessionsResponse["data"];
-  chatSessionsHasMore: boolean;
-  activeChatSessionID: string;
-  activeChatSession: ChatSessionRecord | null;
-  agentChatSessions: AgentChatSessionsResponse["data"];
-  activeAgentChatSessionID: string;
-  activeAgentChatSession: AgentChatSessionRecord | null;
-  requestLedger: RequestLedgerResponse["data"];
-  controlPlaneConfig: ConfiguredStateResponse["data"] | null;
-  retentionRuns: RetentionRunData[];
-  retentionLastRun: RetentionRunData | null;
-  agentAdapterApprovalMode: string;
-};
-
-async function resolveDashboardSnapshot(args: {
-  activeChatSessionID: string;
-  activeAgentChatSessionID: string;
-  previous: DashboardPreviousState;
-}): Promise<DashboardSnapshot> {
-  const results = await loadDashboardResults();
-  const health = requireFulfilledDashboardResult(results.health);
-  const sessionInfo = results.session.status === "fulfilled" ? results.session.value.data : null;
-  const models = resolveModelsResult(results.models);
-  const providers = resolveDashboardResult(results.providers, args.previous.providers);
-  const providerPresets = results.providerPresets.status === "fulfilled" ? results.providerPresets.value.data : [];
-  const agentAdapters = resolveDashboardResult(results.agentAdapters, args.previous.agentAdapters);
-  const budget = resolveDashboardResult(results.budget, args.previous.budget);
-  const accountSummary = resolveDashboardResult(results.accountSummary, args.previous.accountSummary);
-  const requestLedger = resolveDashboardResult(results.requestLedger, args.previous.requestLedger);
-  const controlPlaneConfig = resolveDashboardResult(results.controlPlaneConfig, args.previous.controlPlaneConfig);
-  const retentionRuns = resolveDashboardResult(results.retentionRuns, args.previous.retentionRuns);
-  const retentionLastRun = retentionRuns[0] ?? null;
-  const agentAdapterApprovalMode = results.runtimeStats.status === "fulfilled"
-    ? (results.runtimeStats.value.data.agent_adapter_approval_mode ?? "")
-    : "";
-  const chatState = await resolveChatDashboardState({
-    activeChatSessionID: args.activeChatSessionID,
-    previousSessions: args.previous.chatSessions,
-    previousActiveSession: args.previous.activeChatSession,
-    result: results.chatSessions,
-  });
-  const agentChatState = await resolveAgentChatDashboardState({
-    activeSessionID: args.activeAgentChatSessionID,
-    previousSessions: args.previous.agentChatSessions,
-    previousActiveSession: args.previous.activeAgentChatSession,
-    result: results.agentChatSessions,
-  });
-
-  return {
-    health,
-    sessionInfo,
-    models,
-    providers,
-    providerPresets,
-    agentAdapters,
-    budget,
-    accountSummary,
-    chatSessions: chatState.sessions,
-    chatSessionsHasMore: chatState.hasMore,
-    activeChatSessionID: chatState.activeChatSessionID,
-    activeChatSession: chatState.activeChatSession,
-    agentChatSessions: agentChatState.sessions,
-    activeAgentChatSessionID: agentChatState.activeSessionID,
-    activeAgentChatSession: agentChatState.activeSession,
-    requestLedger,
-    controlPlaneConfig,
-    retentionRuns,
-    retentionLastRun,
-    agentAdapterApprovalMode,
-  };
-}
-
-async function loadDashboardResults(): Promise<DashboardResults> {
-  // Single-user mode: no auth gate. /healthz + /v1/whoami still come
-  // first so the gateway has time to surface its health state, then
-  // every other endpoint fans out in parallel.
-  const [health, session] = await Promise.allSettled([
-    getHealth(),
-    getSession(),
-  ]);
-
-  // Initialize each result as rejected so TS knows these are definitely
-  // assigned before we read them; the inline .then handlers below
-  // overwrite them with the real outcome.
-  const initialReject = <T,>(): PromiseSettledResult<T> => ({ status: "rejected", reason: new Error("uninitialized") });
-  let models: PromiseSettledResult<ModelResponse> = initialReject();
-  let providers: PromiseSettledResult<ProviderStatusResponse> = initialReject();
-  let providerPresets: PromiseSettledResult<{ object: string; data: ProviderPresetRecord[] }> = initialReject();
-  let agentAdapters: PromiseSettledResult<{ object: string; data: AgentAdapterRecord[] }> = initialReject();
-  let budget: PromiseSettledResult<BudgetStatusResponse> = initialReject();
-  let accountSummary: PromiseSettledResult<AccountSummaryResponse> = initialReject();
-  let chatSessions: PromiseSettledResult<ChatSessionsResponse> = initialReject();
-  let agentChatSessions: PromiseSettledResult<AgentChatSessionsResponse> = initialReject();
-  let requestLedger: PromiseSettledResult<RequestLedgerResponse> = initialReject();
-  let controlPlaneConfig: PromiseSettledResult<ConfiguredStateResponse> = initialReject();
-  let retentionRuns: PromiseSettledResult<{ object: string; data: RetentionRunData[] }> = initialReject();
-  let runtimeStats: PromiseSettledResult<RuntimeStatsResponse> = initialReject();
-
-  await Promise.all([
-    getProviderPresets().then(r => { providerPresets = { status: "fulfilled", value: r }; }, e => { providerPresets = { status: "rejected", reason: e }; }),
-    getAgentAdapters().then(r => { agentAdapters = { status: "fulfilled", value: r }; }, e => { agentAdapters = { status: "rejected", reason: e }; }),
-    getChatSessions(20).then(r => { chatSessions = { status: "fulfilled", value: r }; }, e => { chatSessions = { status: "rejected", reason: e }; }),
-    getAgentChatSessions().then(r => { agentChatSessions = { status: "fulfilled", value: r }; }, e => { agentChatSessions = { status: "rejected", reason: e }; }),
-    getModels().then(r => { models = { status: "fulfilled", value: r }; }, e => { models = { status: "rejected", reason: e }; }),
-    getBudget("").then(r => { budget = { status: "fulfilled", value: r }; }, e => { budget = { status: "rejected", reason: e }; }),
-    getAccountSummary("").then(r => { accountSummary = { status: "fulfilled", value: r }; }, e => { accountSummary = { status: "rejected", reason: e }; }),
-    getRequestLedger(20).then(r => { requestLedger = { status: "fulfilled", value: r }; }, e => { requestLedger = { status: "rejected", reason: e }; }),
-    getControlPlaneConfig().then(r => { controlPlaneConfig = { status: "fulfilled", value: r }; }, e => { controlPlaneConfig = { status: "rejected", reason: e }; }),
-    getRetentionRuns(10).then(r => { retentionRuns = { status: "fulfilled", value: r }; }, e => { retentionRuns = { status: "rejected", reason: e }; }),
-    getRuntimeStats().then(r => { runtimeStats = { status: "fulfilled", value: r }; }, e => { runtimeStats = { status: "rejected", reason: e }; }),
-  ]);
-
-  // /admin/providers probes upstream provider runtimes; only call it
-  // when at least one provider has been configured, otherwise the call
-  // returns nothing useful.
-  const configured = controlPlaneConfig.status === "fulfilled" ? (controlPlaneConfig.value.data?.providers ?? []) : [];
-  if (configured.length > 0) {
-    await getProviders().then(
-      r => { providers = { status: "fulfilled", value: r }; },
-      e => { providers = { status: "rejected", reason: e }; },
-    );
-  } else {
-    providers = { status: "fulfilled", value: { object: "list", data: [] } as ProviderStatusResponse };
-  }
-
-  return {
-    health,
-    session,
-    models,
-    providers,
-    providerPresets,
-    agentAdapters,
-    budget,
-    accountSummary,
-    chatSessions,
-    agentChatSessions,
-    requestLedger,
-    controlPlaneConfig,
-    retentionRuns,
-    runtimeStats,
-  };
-}
-
-function requireFulfilledDashboardResult<T>(result: PromiseSettledResult<T>): T {
-  if (result.status === "fulfilled") {
-    return result.value;
-  }
-  throw new Error("failed to load runtime console data");
-}
-
-function resolveModelsResult(result: PromiseSettledResult<ModelResponse>): ModelResponse["data"] {
-  if (result.status === "fulfilled") {
-    return result.value.data;
-  }
-  return [];
-}
-
-function resolveDashboardResult<T>(
-  result: PromiseSettledResult<{ data: T }>,
-  previous: T,
-): T {
-  if (result.status === "fulfilled") {
-    return result.value.data;
-  }
-  return previous;
-}
-
-async function resolveChatDashboardState(args: {
-  activeChatSessionID: string;
-  previousSessions: ChatSessionsResponse["data"];
-  previousActiveSession: ChatSessionRecord | null;
-  result: PromiseSettledResult<ChatSessionsResponse>;
-}): Promise<{
-  sessions: ChatSessionsResponse["data"];
-  hasMore: boolean;
-  activeChatSessionID: string;
-  activeChatSession: ChatSessionRecord | null;
-}> {
-  if (args.result.status !== "fulfilled") {
-    return {
-      sessions: args.previousSessions,
-      hasMore: false,
-      activeChatSessionID: args.activeChatSessionID,
-      activeChatSession: args.previousActiveSession,
-    };
-  }
-
-  const sessions = args.result.value.data ?? [];
-  const hasMore = args.result.value.has_more ?? false;
-  const activeChatSessionID = sessions.some((entry) => entry.id === args.activeChatSessionID)
-    ? args.activeChatSessionID
-    : sessions[0]?.id ?? "";
-
-  if (!activeChatSessionID) {
-    return {
-      sessions,
-      hasMore,
-      activeChatSessionID,
-      activeChatSession: null,
-    };
-  }
-
-  try {
-    const sessionResult = await getChatSession(activeChatSessionID);
-    return {
-      sessions,
-      hasMore,
-      activeChatSessionID,
-      activeChatSession: sessionResult.data,
-    };
-  } catch {
-    return {
-      sessions,
-      hasMore,
-      activeChatSessionID,
-      activeChatSession: null,
-    };
-  }
-}
-
-async function resolveAgentChatDashboardState(args: {
-  activeSessionID: string;
-  previousSessions: AgentChatSessionsResponse["data"];
-  previousActiveSession: AgentChatSessionRecord | null;
-  result: PromiseSettledResult<AgentChatSessionsResponse>;
-}): Promise<{
-  sessions: AgentChatSessionsResponse["data"];
-  activeSessionID: string;
-  activeSession: AgentChatSessionRecord | null;
-}> {
-  if (args.result.status !== "fulfilled") {
-    return {
-      sessions: args.previousSessions,
-      activeSessionID: args.activeSessionID,
-      activeSession: args.previousActiveSession,
-    };
-  }
-
-  const sessions = args.result.value.data ?? [];
-  const activeSessionID = args.activeSessionID && sessions.some((entry) => entry.id === args.activeSessionID)
-    ? args.activeSessionID
-    : "";
-
-  if (!activeSessionID) {
-    return { sessions, activeSessionID, activeSession: null };
-  }
-
-  try {
-    const sessionResult = await getAgentChatSession(activeSessionID);
-    return { sessions, activeSessionID, activeSession: sessionResult.data };
-  } catch (error) {
-    if (!(error instanceof ApiError) || error.status !== 404) {
-      return { sessions, activeSessionID, activeSession: args.previousActiveSession };
-    }
-    return { sessions, activeSessionID: "", activeSession: null };
-  }
-}
-
-// Single-user mode: the session label is fixed. The whoami endpoint
-// is still called to surface gateway features in the future, but we
-// don't read role/tenant/auth from it.
-function deriveSessionState(_sessionInfo: SessionResponse["data"] | null): SessionState {
-  return { label: "Local" };
-}
