+++ ui/src/features/chats/ChatView.tsx
@@ -2,11 +2,11 @@ import { useEffect, useRef, useState } from "react";
 import type { SyntheticEvent } from "react";
 import type { RuntimeConsoleViewModel } from "../../app/useRuntimeConsole";
 import { describeGatewayError, formatErrorCode } from "../../lib/error-diagnostics";
-import { parseInlineNodes, parseMarkdownBlocks } from "../../lib/markdown";
-import type { AgentAdapterRecord, AgentChatActivityRecord, AgentChatChangedFileDiffRecord, AgentChatChangedFileRecord, AgentChatSessionRecord, AgentChatUsageRecord } from "../../types/runtime";
+import type { AgentAdapterRecord, AgentChatActivityRecord, AgentChatSessionRecord, AgentChatUsageRecord } from "../../types/runtime";
 import { AgentAdapterPicker, CodeBlock, Icon, Icons, InlineError, ModelPicker, ProviderPicker } from "../shared/ui";
 import { AgentApprovalAutoModeBanner, AgentApprovalsBanner } from "./AgentApprovalBanner";
 import { AgentApprovalModal } from "./AgentApprovalModal";
+import { MessageRow } from "./MessageRow";
 
 type Props = {
   state: RuntimeConsoleViewModel["state"];
@@ -1149,786 +1149,4 @@ function formatDuration(durationMS: number): string {
   const minutes = Math.floor(seconds / 60);
   const rest = Math.round(seconds % 60);
   return `${minutes}m ${rest}s`;
-}
-
-// (ModelPicker now lives in shared/ui — single component shared by the
-// chat header, the new-task slideover, and any future surface that
-// needs to pick a model with type-to-filter + disabled-provider
-// awareness.)
-
-function MessageRow({ id, role, model, content, time, promptTokens, completionTokens, costUsd, badge, runtimeMeta, activities, diffStat, diff, agentSessionID, onListAgentFiles, onGetAgentFileDiff, onRevertAgentFiles, rawOutput, agentUsage, error, onCopy, copied }: {
-  id: string; role: "user" | "assistant"; model?: string; content: string;
-  time: string; promptTokens?: number; completionTokens?: number; costUsd?: string;
-  badge?: string; runtimeMeta?: string;
-  activities?: AgentChatActivityRecord[]; diffStat?: string; diff?: string; agentSessionID?: string;
-  onListAgentFiles?: (sessionID: string, messageID: string) => Promise<AgentChatChangedFileRecord[]>;
-  onGetAgentFileDiff?: (sessionID: string, messageID: string, path: string) => Promise<AgentChatChangedFileDiffRecord | null>;
-  onRevertAgentFiles?: (sessionID: string, messageID: string, paths: string[]) => Promise<boolean>;
-  rawOutput?: string; agentUsage?: AgentChatUsageRecord; error?: string;
-  onCopy: (id: string, text: string) => void; copied: boolean;
-}) {
-  const [hovered, setHovered] = useState(false);
-  const isAssistant = role === "assistant";
-  const hasTokenData = isAssistant && (promptTokens ?? 0) > 0;
-  const showRawOutput = isAssistant && rawOutput && rawOutput.trim() && rawOutput.trim() !== content.trim();
-  const waitingForAgentOutput = isAssistant && !content.trim() && activities?.some(activity => activity.status === "running");
-  const failed = isAssistant && badge === "failed";
-  const cancelled = isAssistant && badge === "cancelled";
-  const thinkingForAgent = isAssistant
-    && badge === "running"
-    && content.trim() !== ""
-    && isLikelyTransientAgentNarration(content)
-    && !(activities ?? []).some(activity => activity.type === "tool_call");
-
-  return (
-    <div onMouseEnter={() => setHovered(true)} onMouseLeave={() => setHovered(false)}
-      style={{ padding: "4px 16px 12px", maxWidth: 820, margin: "0 auto", width: "100%" }}>
-      <div style={{ display: "flex", gap: 10, alignItems: "flex-start" }}>
-        <div style={{
-          width: 28, height: 28, borderRadius: "var(--radius-sm)", flexShrink: 0, marginTop: 2,
-          background: isAssistant ? "var(--teal-bg)" : "var(--bg3)",
-          border: `1px solid ${isAssistant ? "var(--teal-border)" : "var(--border)"}`,
-          display: "flex", alignItems: "center", justifyContent: "center",
-        }}>
-          <span style={{ fontFamily: "var(--font-mono)", fontSize: 11, color: isAssistant ? "var(--teal)" : "var(--t1)", fontWeight: 600 }}>
-            {isAssistant ? (model || "H")[0].toUpperCase() : "U"}
-          </span>
-        </div>
-        <div style={{ flex: 1, minWidth: 0 }}>
-          <div style={{ display: "flex", alignItems: "center", gap: 8, marginBottom: 5 }}>
-            {isAssistant
-              ? <span style={{ fontFamily: "var(--font-mono)", fontSize: 11, color: "var(--teal)" }}>{model || "hecate"}</span>
-              : <span style={{ fontSize: 11, color: "var(--t2)", fontWeight: 500 }}>You</span>
-            }
-            <span style={{ fontSize: 10, color: "var(--t3)", fontFamily: "var(--font-mono)" }}>{time}</span>
-            {hasTokenData && (
-              <span style={{ fontSize: 10, color: "var(--t3)", fontFamily: "var(--font-mono)" }}>
-                {promptTokens}↑ {completionTokens}↓
-                {costUsd && costUsd !== "0" ? ` · $${Number(costUsd).toFixed(5)}` : ""}
-              </span>
-            )}
-            {isAssistant && badge && (
-              <span className="badge badge-muted" style={{ fontSize: 10 }}>{badge}</span>
-            )}
-            {isAssistant && runtimeMeta && (
-              <span style={{ fontSize: 10, color: "var(--t3)", fontFamily: "var(--font-mono)" }}>{runtimeMeta}</span>
-            )}
-            <div style={{ marginLeft: "auto", display: "flex", gap: 4, opacity: hovered ? 1 : 0, transition: "opacity 0.15s" }}>
-              <button className="btn btn-ghost btn-sm" style={{ padding: "2px 6px", gap: 4 }}
-                onClick={() => onCopy(id, content)}>
-                <Icon d={copied ? Icons.check : Icons.copy} size={12} />
-              </button>
-            </div>
-          </div>
-          {failed || cancelled ? (
-            <AgentRunNotice status={failed ? "failed" : "cancelled"} message={error || content} />
-          ) : thinkingForAgent ? (
-            <AgentLiveText content={content} />
-          ) : waitingForAgentOutput ? (
-            <div style={{ alignItems: "center", color: "var(--t2)", display: "flex", fontSize: 13, gap: 8, lineHeight: 1.7 }}>
-              <span style={{ background: "var(--teal)", borderRadius: 999, display: "inline-block", height: 6, opacity: 0.8, width: 6 }} />
-              Waiting for agent output...
-            </div>
-          ) : (
-            <Markdown content={content} />
-          )}
-          {isAssistant && activities && activities.length > 0 && (
-            <ActivityTimeline activities={activities} diffStat={diffStat} />
-          )}
-          {isAssistant && agentUsage && !agentUsageEmpty(agentUsage) && (
-            <AgentUsage usage={agentUsage} />
-          )}
-          {isAssistant && (diff || diffStat) && (
-            <AgentDiffReview
-              sessionID={agentSessionID ?? ""}
-              messageID={id}
-              diffStat={diffStat}
-              diff={diff}
-              onListFiles={onListAgentFiles}
-              onGetFileDiff={onGetAgentFileDiff}
-              onRevertFiles={onRevertAgentFiles}
-            />
-          )}
-          {showRawOutput && (
-            <details style={{ marginTop: 8 }}>
-              <summary style={{ cursor: "pointer", fontFamily: "var(--font-mono)", fontSize: 11, color: "var(--t3)" }}>
-                raw adapter output{rawOutput ? ` · ${formatLineCount(rawOutput)}` : ""}
-              </summary>
-              <div style={{ marginTop: 6 }}>
-                <CodeBlock code={rawOutput} lang="text" />
-              </div>
-            </details>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function isLikelyTransientAgentNarration(text: string): boolean {
-  const normalized = text.trim().toLowerCase();
-  if (!normalized) return false;
-  return [
-    "i'll ",
-    "i’ll ",
-    "i will ",
-    "i'm going to ",
-    "i’m going to ",
-    "i'm checking ",
-    "i’m checking ",
-    "i'll check ",
-    "i’ll check ",
-    "i'll inspect ",
-    "i’ll inspect ",
-    "let me ",
-    "checking ",
-  ].some(prefix => normalized.startsWith(prefix));
-}
-
-function AgentRunNotice({ status, message }: { status: "failed" | "cancelled"; message: string }) {
-  const color = status === "failed" ? "var(--red)" : "var(--amber)";
-  return (
-    <div style={{
-      border: "1px solid var(--border)",
-      borderLeft: `3px solid ${color}`,
-      borderRadius: "var(--radius-sm)",
-      background: "var(--bg2)",
-      padding: "9px 10px",
-    }}>
-      <div style={{ color, fontFamily: "var(--font-mono)", fontSize: 11, marginBottom: 4 }}>
-        agent run {status}
-      </div>
-      {message && (
-        <div style={{ color: "var(--t0)", fontSize: 13, lineHeight: 1.6, whiteSpace: "pre-wrap" }}>
-          {message}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function AgentLiveText({ content }: { content: string }) {
-  return (
-    <div style={{ alignItems: "baseline", display: "flex", gap: 6, minWidth: 0 }}>
-      <div style={{ color: "var(--t0)", flex: "0 1 auto", fontSize: 13, lineHeight: 1.7, minWidth: 0, whiteSpace: "pre-wrap" }}>
-        {content}
-      </div>
-      <span
-        aria-hidden="true"
-        style={{
-          animation: "hecate-live-caret 1.1s ease-in-out infinite",
-          background: "var(--teal)",
-          borderRadius: 999,
-          display: "inline-block",
-          flexShrink: 0,
-          height: 5,
-          opacity: 0.75,
-          transform: "translateY(-1px)",
-          width: 5,
-        }}
-      />
-    </div>
-  );
-}
-
-function AgentUsage({ usage }: { usage: AgentChatUsageRecord }) {
-  const cost = formatAgentReportedCost(usage);
-  const context = formatAgentContextUsage(usage);
-  return (
-    <div style={{ display: "flex", flexWrap: "wrap", gap: 8, marginTop: 8, fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-      {cost && <span>{cost}</span>}
-      {context && <span>{context}</span>}
-      <span>reported by adapter · not enforced by Hecate</span>
-    </div>
-  );
-}
-
-function agentUsageEmpty(usage: AgentChatUsageRecord): boolean {
-  return !usage.reported_cost_amount && !usage.reported_cost_currency && !(usage.context_size ?? 0) && !(usage.context_used ?? 0);
-}
-
-function formatAgentReportedCost(usage: AgentChatUsageRecord): string {
-  if (!usage.reported_cost_amount && !usage.reported_cost_currency) return "";
-  const currency = usage.reported_cost_currency ? ` ${usage.reported_cost_currency}` : "";
-  return `${usage.reported_cost_amount || "0"}${currency}`;
-}
-
-function formatAgentContextUsage(usage: AgentChatUsageRecord): string {
-  const used = usage.context_used ?? 0;
-  const size = usage.context_size ?? 0;
-  if (!used && !size) return "";
-  if (!size) return `${used} context used`;
-  return `${used}/${size} context`;
-}
-
-function AgentDiffReview({
-  sessionID,
-  messageID,
-  diffStat,
-  diff,
-  onListFiles,
-  onGetFileDiff,
-  onRevertFiles,
-}: {
-  sessionID: string;
-  messageID: string;
-  diffStat?: string;
-  diff?: string;
-  onListFiles?: (sessionID: string, messageID: string) => Promise<AgentChatChangedFileRecord[]>;
-  onGetFileDiff?: (sessionID: string, messageID: string, path: string) => Promise<AgentChatChangedFileDiffRecord | null>;
-  onRevertFiles?: (sessionID: string, messageID: string, paths: string[]) => Promise<boolean>;
-}) {
-  const [files, setFiles] = useState<AgentChatChangedFileRecord[] | null>(null);
-  const [selectedDiff, setSelectedDiff] = useState<AgentChatChangedFileDiffRecord | null>(null);
-  const [loadingFiles, setLoadingFiles] = useState(false);
-  const [loadingPath, setLoadingPath] = useState("");
-  const [revertingPath, setRevertingPath] = useState("");
-  const [confirmRevertPath, setConfirmRevertPath] = useState("");
-  const [localError, setLocalError] = useState("");
-  const hasReviewAPI = Boolean(sessionID && onListFiles && onGetFileDiff && onRevertFiles);
-
-  async function loadFiles() {
-    if (!hasReviewAPI || !onListFiles) return;
-    setLoadingFiles(true);
-    setLocalError("");
-    try {
-      const nextFiles = await onListFiles(sessionID, messageID);
-      setFiles(nextFiles);
-    } catch {
-      setLocalError("Could not load changed files. The captured diff may no longer be available.");
-    } finally {
-      setLoadingFiles(false);
-    }
-  }
-
-  async function inspectFile(file: AgentChatChangedFileRecord) {
-    if (!hasReviewAPI || !onGetFileDiff) return;
-    setLoadingPath(file.path);
-    setLocalError("");
-    try {
-      const nextDiff = await onGetFileDiff(sessionID, messageID, file.path);
-      if (nextDiff) {
-        setSelectedDiff(nextDiff);
-      } else {
-        setLocalError("Could not load that file diff.");
-      }
-    } catch {
-      setLocalError("Could not load that file diff.");
-    } finally {
-      setLoadingPath("");
-    }
-  }
-
-  async function confirmRevert(paths: string[], label: string) {
-    if (!hasReviewAPI || !onRevertFiles) return;
-    setRevertingPath(label);
-    setLocalError("");
-    try {
-      const ok = await onRevertFiles(sessionID, messageID, paths);
-      if (!ok) {
-        setLocalError("Revert failed. The workspace may not be a Git repository, or the file changed since capture.");
-        return;
-      }
-      setSelectedDiff(null);
-      setFiles(null);
-      await loadFiles();
-    } catch {
-      setLocalError("Revert failed. The workspace may not be a Git repository, or the file changed since capture.");
-    } finally {
-      setRevertingPath("");
-      setConfirmRevertPath("");
-    }
-  }
-
-  const summary = diffStat ? formatDiffStatSummary(diffStat) : "";
-  const visibleFiles = files ?? [];
-
-  return (
-    <details
-      data-testid="agent-diff-review"
-      onToggle={(event: SyntheticEvent<HTMLDetailsElement>) => {
-        if (event.currentTarget.open && files === null && !loadingFiles) {
-          void loadFiles();
-        }
-      }}
-      style={{ marginTop: 8 }}
-    >
-      <summary style={{ cursor: "pointer", fontFamily: "var(--font-mono)", fontSize: 11, color: "var(--t3)" }}>
-        files changed{summary ? ` · ${summary}` : ""}
-      </summary>
-      <div style={{ display: "grid", gap: 8, marginTop: 6 }}>
-        <div style={{ color: "var(--t3)", fontSize: 11, lineHeight: 1.5 }}>
-          External-agent changes are already in your workspace. Inspect the captured Git diff, keep it, or revert selected paths.
-        </div>
-        {!hasReviewAPI && (
-          <>
-            {diffStat && <DiffStatList diffStat={diffStat} />}
-            {diff && <CodeBlock code={diff} lang="diff" />}
-          </>
-        )}
-        {hasReviewAPI && (
-          <>
-            {loadingFiles && (
-              <div style={{ color: "var(--t2)", fontFamily: "var(--font-mono)", fontSize: 11 }}>Loading changed files...</div>
-            )}
-            {!loadingFiles && visibleFiles.length === 0 && diffStat && (
-              <DiffStatList diffStat={diffStat} />
-            )}
-            {visibleFiles.length > 0 && (
-              <div style={{
-                border: "1px solid var(--border)",
-                borderRadius: "var(--radius-sm)",
-                background: "var(--bg2)",
-                overflow: "hidden",
-              }}>
-                <div style={{
-                  alignItems: "center",
-                  borderBottom: "1px solid var(--border)",
-                  display: "flex",
-                  gap: 8,
-                  justifyContent: "space-between",
-                  padding: "7px 9px",
-                }}>
-                  <span style={{ color: "var(--t2)", fontFamily: "var(--font-mono)", fontSize: 11 }}>
-                    {visibleFiles.length} changed file{visibleFiles.length === 1 ? "" : "s"}
-                  </span>
-                  {confirmRevertPath === "__all__" ? (
-                    <div style={{ display: "flex", gap: 6 }}>
-                      <button
-                        className="btn btn-ghost btn-sm"
-                        disabled={Boolean(revertingPath)}
-                        onClick={() => void confirmRevert([], "__all__")}
-                        type="button"
-                      >
-                        {revertingPath === "__all__" ? "Reverting..." : "Confirm revert all"}
-                      </button>
-                      <button className="btn btn-ghost btn-sm" onClick={() => setConfirmRevertPath("")} type="button">Cancel</button>
-                    </div>
-                  ) : (
-                    <button
-                      className="btn btn-ghost btn-sm"
-                      disabled={Boolean(revertingPath)}
-                      onClick={() => setConfirmRevertPath("__all__")}
-                      type="button"
-                    >
-                      Revert all
-                    </button>
-                  )}
-                </div>
-                <div style={{ display: "grid" }}>
-                  {visibleFiles.map(file => (
-                    <div
-                      key={file.path}
-                      style={{
-                        alignItems: "center",
-                        borderTop: "1px solid var(--border)",
-                        display: "grid",
-                        gap: 8,
-                        gridTemplateColumns: "minmax(0, 1fr) auto",
-                        padding: "7px 9px",
-                      }}
-                    >
-                      <div style={{ minWidth: 0 }}>
-                        <div style={{ color: "var(--t1)", fontFamily: "var(--font-mono)", fontSize: 11, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>
-                          {file.path}
-                        </div>
-                        <div style={{ color: "var(--t3)", fontFamily: "var(--font-mono)", fontSize: 10, marginTop: 2 }}>
-                          {formatChangedFileMeta(file)}
-                        </div>
-                      </div>
-                      {confirmRevertPath === file.path ? (
-                        <div style={{ display: "flex", gap: 6 }}>
-                          <button
-                            className="btn btn-ghost btn-sm"
-                            disabled={Boolean(revertingPath)}
-                            onClick={() => void confirmRevert([file.path], file.path)}
-                            type="button"
-                          >
-                            {revertingPath === file.path ? "Reverting..." : `Confirm revert ${file.path}`}
-                          </button>
-                          <button className="btn btn-ghost btn-sm" onClick={() => setConfirmRevertPath("")} type="button">Cancel</button>
-                        </div>
-                      ) : (
-                        <div style={{ display: "flex", gap: 6 }}>
-                          <button
-                            className="btn btn-ghost btn-sm"
-                            disabled={loadingPath === file.path || Boolean(revertingPath)}
-                            onClick={() => void inspectFile(file)}
-                            type="button"
-                          >
-                            {loadingPath === file.path ? "Loading..." : `Inspect ${file.path}`}
-                          </button>
-                          <button
-                            className="btn btn-ghost btn-sm"
-                            disabled={Boolean(revertingPath)}
-                            onClick={() => setConfirmRevertPath(file.path)}
-                            type="button"
-                          >
-                            Revert {file.path}
-                          </button>
-                        </div>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )}
-            {localError && <InlineError message={localError} />}
-            {selectedDiff && (
-              <div style={{ display: "grid", gap: 6 }}>
-                <div style={{ color: "var(--t2)", fontFamily: "var(--font-mono)", fontSize: 11 }}>
-                  diff · {selectedDiff.path}
-                </div>
-                <CodeBlock code={selectedDiff.diff} lang="diff" />
-              </div>
-            )}
-          </>
-        )}
-      </div>
-    </details>
-  );
-}
-
-function formatChangedFileMeta(file: AgentChatChangedFileRecord): string {
-  const parts = [file.status || "modified"];
-  if (file.additions > 0) parts.push(`+${file.additions}`);
-  if (file.deletions > 0) parts.push(`-${file.deletions}`);
-  if (parts.length === 1) parts.push("no line delta");
-  return parts.join(" · ");
-}
-
-function DiffStatList({ diffStat }: { diffStat: string }) {
-  const rows = parseDiffStatRows(diffStat);
-  const summary = formatDiffStatSummary(diffStat);
-
-  if (rows.length === 0) {
-    return (
-      <div style={{ color: "var(--t2)", fontFamily: "var(--font-mono)", fontSize: 11 }}>
-        {summary}
-      </div>
-    );
-  }
-
-  return (
-    <div style={{
-      display: "grid",
-      gap: 5,
-      padding: "8px 10px",
-      border: "1px solid var(--border)",
-      borderRadius: "var(--radius-sm)",
-      background: "var(--bg2)",
-    }}>
-      {rows.map(row => (
-        <div key={row.path} style={{ display: "grid", gridTemplateColumns: "minmax(0, 1fr) auto", gap: 10, alignItems: "baseline" }}>
-          <span style={{ color: "var(--t1)", fontFamily: "var(--font-mono)", fontSize: 11, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>
-            {row.path}
-          </span>
-          <span style={{ color: "var(--t3)", fontFamily: "var(--font-mono)", fontSize: 11, whiteSpace: "nowrap" }}>
-            {row.change}
-          </span>
-        </div>
-      ))}
-      {summary && (
-        <div style={{ borderTop: "1px solid var(--border)", color: "var(--t2)", fontFamily: "var(--font-mono)", fontSize: 11, marginTop: 2, paddingTop: 6 }}>
-          {summary}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function parseDiffStatRows(diffStat: string): Array<{ path: string; change: string }> {
-  return diffStat
-    .split(/\r?\n/)
-    .map(line => line.trim())
-    .filter(Boolean)
-    .filter(line => !/\bfiles? changed\b/.test(line))
-    .map(line => {
-      const match = line.match(/^(.+?)\s+\|\s+(.+)$/);
-      if (!match) return null;
-      return { path: match[1].trim(), change: match[2].trim() };
-    })
-    .filter((row): row is { path: string; change: string } => row !== null);
-}
-
-function formatDiffStatSummary(diffStat: string): string {
-  const lines = diffStat.split(/\r?\n/).map(line => line.trim()).filter(Boolean);
-  return lines.find(line => /\bfiles? changed\b/.test(line)) || lines[0] || "";
-}
-
-function ActivityTimeline({ activities, diffStat }: { activities: AgentChatActivityRecord[]; diffStat?: string }) {
-  const visible = compactAgentActivities(activities);
-  if (visible.length === 0) return null;
-  const terminal = terminalAgentActivity(activities);
-  const hasRunning = !terminal && activities.some(activity => activity.status === "running");
-  const plan = visible.filter(activity => activity.type === "plan");
-  const tools = visible.filter(activity => activity.type === "tool_call");
-  const other = visible.filter(activity => activity.type !== "plan" && activity.type !== "tool_call");
-  const summary = [
-    terminal ? terminal.status : hasRunning ? "running" : "details",
-    plan.length > 0 ? `${plan.filter(item => item.status === "completed").length}/${plan.length} plan` : "",
-    tools.length > 0 ? `${tools.length} tool${tools.length === 1 ? "" : "s"}` : "",
-    diffStat ? "files changed" : "",
-  ].filter(Boolean).join(" · ");
-
-  return (
-    <details open={hasRunning} style={{ marginTop: 8 }}>
-      <summary style={{ cursor: "pointer", fontFamily: "var(--font-mono)", fontSize: 11, color: "var(--t3)" }}>
-        {summary}
-      </summary>
-      <div style={{
-        display: "grid",
-        gap: 5,
-        marginTop: 6,
-        padding: "8px 10px",
-        border: "1px solid var(--border)",
-        borderRadius: "var(--radius-sm)",
-        background: "var(--bg2)",
-      }}>
-        {plan.length > 0 && <PlanActivityList items={plan} />}
-        {tools.length > 0 && <ToolActivityList items={tools} />}
-        {other.map((activity, index) => (
-          <ActivityLine key={activity.id || `${activity.type}-${activity.created_at ?? index}`} activity={activity} />
-        ))}
-      </div>
-    </details>
-  );
-}
-
-function PlanActivityList({ items }: { items: AgentChatActivityRecord[] }) {
-  return (
-    <div style={{ display: "grid", gap: 5 }}>
-      {items.map((activity, index) => (
-        <div key={activity.id || `${activity.title}-${index}`} style={{ display: "flex", alignItems: "baseline", gap: 8, minWidth: 0 }}>
-          <span style={{ color: activity.status === "completed" ? "var(--green)" : activity.status === "in_progress" ? "var(--teal)" : "var(--t3)", flexShrink: 0, fontFamily: "var(--font-mono)", fontSize: 11 }}>
-            {activity.status === "completed" ? "x" : activity.status === "in_progress" ? ">" : "-"}
-          </span>
-          <span style={{ color: "var(--t1)", fontSize: 11, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>
-            {activity.title}
-          </span>
-          {activity.kind && (
-            <span style={{ color: "var(--t3)", flexShrink: 0, fontFamily: "var(--font-mono)", fontSize: 10 }}>
-              {activity.kind}
-            </span>
-          )}
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function ToolActivityList({ items }: { items: AgentChatActivityRecord[] }) {
-  return (
-    <div style={{ display: "grid", gap: 5 }}>
-      {items.map((activity, index) => (
-        <ActivityLine key={activity.id || `${activity.type}-${activity.created_at ?? index}`} activity={activity} prefix={activity.kind || "tool"} />
-      ))}
-    </div>
-  );
-}
-
-function ActivityLine({ activity, prefix }: { activity: AgentChatActivityRecord; prefix?: string }) {
-  return (
-    <div style={{ display: "flex", alignItems: "baseline", gap: 8, minWidth: 0 }}>
-      <span style={{
-        width: 7,
-        height: 7,
-        borderRadius: 999,
-        background: activityStatusColor(activity.status),
-        flexShrink: 0,
-      }} />
-      {prefix && (
-        <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)", whiteSpace: "nowrap" }}>
-          {prefix}
-        </span>
-      )}
-      <span style={{ fontFamily: "var(--font-mono)", fontSize: 11, color: "var(--t1)", whiteSpace: "nowrap" }}>
-        {activity.title}
-      </span>
-      {activity.detail && (
-        <span style={{ fontSize: 11, color: "var(--t3)", overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>
-          {activity.detail}
-        </span>
-      )}
-    </div>
-  );
-}
-
-function compactAgentActivities(activities: AgentChatActivityRecord[]): AgentChatActivityRecord[] {
-  const hiddenTypes = new Set(["output"]);
-  const terminal = terminalAgentActivity(activities);
-  const out: AgentChatActivityRecord[] = [];
-  for (const activity of activities) {
-    if (hiddenTypes.has(activity.type)) continue;
-    if (activity.type === "completed" && activity.title.toLowerCase() === "final answer") continue;
-    if (terminal && (activity.type === "started" || activity.type === "running")) continue;
-    if (activity.type === "running" && activities.some(item => item.type === "output")) continue;
-    out.push(activity);
-  }
-  return out;
-}
-
-function terminalAgentActivity(activities: AgentChatActivityRecord[]): AgentChatActivityRecord | undefined {
-  const terminalTypes = new Set(["completed", "failed", "cancelled"]);
-  return [...activities].reverse().find(activity => terminalTypes.has(activity.type));
-}
-
-function activityStatusColor(status?: string) {
-  switch (status) {
-  case "failed":
-    return "var(--red)";
-  case "cancelled":
-    return "var(--amber)";
-  case "running":
-    return "var(--teal)";
-  default:
-    return "var(--green)";
-  }
-}
-
-function formatLineCount(value: string): string {
-  const trimmed = value.trim();
-  if (!trimmed) return "0 lines";
-  const count = trimmed.split(/\r?\n/).length;
-  return `${count} line${count === 1 ? "" : "s"}`;
-}
-
-function Markdown({ content }: { content: string }) {
-  const blocks = parseMarkdownBlocks(content);
-  return (
-    <div style={{ fontSize: 13, color: "var(--t0)", lineHeight: 1.7 }}>
-      {blocks.map((block, i) => {
-        if (block.type === "code") {
-          return <CodeBlock key={i} code={block.text} lang={block.lang ?? ""} />;
-        }
-        if (block.type === "heading") {
-          const sizes: Record<number, string> = { 1: "16px", 2: "14px", 3: "13px" };
-          return (
-            <div key={i} style={{ fontWeight: 600, fontSize: sizes[block.level ?? 1] ?? "13px", margin: "10px 0 4px", color: "var(--t0)" }}>
-              {renderInline(block.text)}
-            </div>
-          );
-        }
-        if (block.type === "ul") {
-          return (
-            <ul key={i} style={{ margin: "4px 0 4px 20px", padding: 0 }}>
-              {block.items!.map((item, j) => (
-                <li key={j} style={{ marginBottom: 2 }}>{renderInline(item)}</li>
-              ))}
-            </ul>
-          );
-        }
-        if (block.type === "ol") {
-          return (
-            <ol key={i} style={{ margin: "4px 0 4px 20px", padding: 0 }}>
-              {block.items!.map((item, j) => (
-                <li key={j} style={{ marginBottom: 2 }}>{renderInline(item)}</li>
-              ))}
-            </ol>
-          );
-        }
-        if (block.type === "task") {
-          return (
-            <ul key={i} style={{ display: "grid", gap: 4, listStyle: "none", margin: "6px 0", padding: 0 }}>
-              {block.tasks!.map((task, j) => (
-                <li key={j} style={{ alignItems: "flex-start", display: "flex", gap: 8 }}>
-                  <span
-                    aria-label={task.checked ? "Completed task" : "Incomplete task"}
-                    role="img"
-                    style={{
-                      alignItems: "center",
-                      background: task.checked ? "var(--teal-soft)" : "var(--bg3)",
-                      border: `1px solid ${task.checked ? "var(--teal-border)" : "var(--border)"}`,
-                      borderRadius: 4,
-                      color: task.checked ? "var(--teal)" : "transparent",
-                      display: "inline-flex",
-                      flex: "0 0 auto",
-                      fontSize: 10,
-                      height: 14,
-                      justifyContent: "center",
-                      marginTop: 4,
-                      width: 14,
-                    }}
-                  >
-                    x
-                  </span>
-                  <span>{renderInline(task.text)}</span>
-                </li>
-              ))}
-            </ul>
-          );
-        }
-        if (block.type === "table") {
-          return (
-            <div key={i} style={{ border: "1px solid var(--border)", borderRadius: "var(--radius-sm)", margin: "8px 0", overflowX: "auto" }}>
-              <table style={{ borderCollapse: "collapse", minWidth: "100%", fontSize: 12 }}>
-                <thead>
-                  <tr>
-                    {block.table!.headers.map((header, j) => (
-                      <th key={j} style={{ background: "var(--bg3)", borderBottom: "1px solid var(--border)", color: "var(--t1)", fontWeight: 600, padding: "6px 8px", textAlign: "left", whiteSpace: "nowrap" }}>
-                        {renderInline(header)}
-                      </th>
-                    ))}
-                  </tr>
-                </thead>
-                <tbody>
-                  {block.table!.rows.map((row, j) => (
-                    <tr key={j}>
-                      {block.table!.headers.map((_, k) => (
-                        <td key={k} style={{ borderTop: j === 0 ? "none" : "1px solid var(--border)", color: "var(--t0)", padding: "6px 8px", verticalAlign: "top" }}>
-                          {renderInline(row[k] ?? "")}
-                        </td>
-                      ))}
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          );
-        }
-        if (block.type === "hr") {
-          return <hr key={i} style={{ border: "none", borderTop: "1px solid var(--border)", margin: "10px 0" }} />;
-        }
-        // paragraph
-        return (
-          <p key={i} style={{ margin: "0 0 6px", whiteSpace: "pre-wrap" }}>
-            {renderInline(block.text)}
-          </p>
-        );
-      })}
-    </div>
-  );
-}
-
-function renderInline(text: string): React.ReactNode {
-  return parseInlineNodes(text).map((node, i) => {
-    if (node.t === "bold") return <strong key={i}>{node.v}</strong>;
-    if (node.t === "italic") return <em key={i}>{node.v}</em>;
-    if (node.t === "code") return (
-      <code key={i} style={{ fontFamily: "var(--font-mono)", fontSize: "0.9em", background: "var(--bg3)", padding: "1px 4px", borderRadius: "var(--radius-sm)", color: "var(--teal)" }}>
-        {node.v}
-      </code>
-    );
-    if (node.t === "link") {
-      return (
-        <a
-          key={i}
-          href={safeMarkdownHref(node.href)}
-          rel="noreferrer"
-          target="_blank"
-          style={{ color: "var(--teal)", textDecoration: "none", borderBottom: "1px solid var(--teal-border)" }}
-        >
-          {node.v}
-        </a>
-      );
-    }
-    return node.v;
-  });
-}
-
-function safeMarkdownHref(href: string): string {
-  if (/^https?:\/\//i.test(href) || /^mailto:/i.test(href)) {
-    return href;
-  }
-  return "#";
 }
